allow floating point values for scale

+++ include/gint-extras/display.h
@@ -3,7 +3,7 @@
 
 #include <gint/display.h>
 
-void dimage_scale(int, int, int, const bopti_image_t*);
-void dsubimage_scale(int, int, const bopti_image_t*, int, int, int, int, int);
+void dimage_scale(int, int, float, const bopti_image_t*);
+void dsubimage_scale(int, int, const bopti_image_t*, int, int, int, int, float);
 
 #endif
