+++ src/gint-extras/display.c
@@ -4,7 +4,7 @@
 
 typedef struct image_linear_map image_linear_map;
 
-void dimage_scale(int x, int y, int scale, const bopti_image_t *image) {
+void dimage_scale(int x, int y, float scale, const bopti_image_t *image) {
 	image_linear_map map;
 	image_scale(image, scale*65536, scale*65536, &map);
 
@@ -14,7 +14,7 @@ void dimage_scale(int x, int y, int scale, const bopti_image_t *image) {
 	image_free(scaled_image);
 }
 
-void dsubimage_scale(int x, int y, const bopti_image_t *image, int left, int top, int width, int height, int scale) {
+void dsubimage_scale(int x, int y, const bopti_image_t *image, int left, int top, int width, int height, float scale) {
 	bopti_image_t sub_image;
 	image_sub(image, left, top, width, height, &sub_image);
 	dimage_scale(x, y, scale, &sub_image);
